Refactor controllers
* Delete controllers/index
* Move more launch logic into app controller
