+++ packages/neuron-wallet/src/controllers/api.ts
@@ -4,11 +4,16 @@ import { ipcMain, IpcMainInvokeEvent } from 'electron'
 import env from 'env'
 import i18n from 'utils/i18n'
 import { showWindow } from './app/show-window'
-import { TransactionsController, WalletsController, SyncController, NetworksController, UpdateController, DaoController } from 'controllers'
+import TransactionsController from 'controllers/transactions'
+import WalletsController from 'controllers/wallets'
+import SyncController from 'controllers/sync'
+import NetworksController from 'controllers/networks'
+import UpdateController from 'controllers/update'
+import DaoController from 'controllers/dao'
 import { NetworkType, Network } from 'models/network'
+import { ConnectionStatusSubject } from 'models/subjects/node'
 import NetworksService from 'services/networks'
 import WalletsService from 'services/wallets'
-import { ConnectionStatusSubject } from 'models/subjects/node'
 import { ResponseCode } from 'utils/const'
 import { TransactionWithoutHash, OutPoint } from 'types/cell-types'
