--- packages/neuron-wallet/src/controllers/index.ts
@@ -1,20 +1,0 @@
-import AppController from './app'
-import NetworksController from './networks'
-import WalletsController from './wallets'
-import TransactionsController from './transactions'
-import SyncController from './sync'
-import UpdateController from './update'
-import DaoController from './dao'
-
-import ApiController from './api'
-
-export {
-  AppController,
-  NetworksController,
-  WalletsController,
-  TransactionsController,
-  SyncController,
-  UpdateController,
-  DaoController,
-  ApiController,
-}
